+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [1.4.0] - 2020-07-09
+### Added
+* Active config values are now shown as part of verbose output
+* File extension can now be set with `format` in config and/or at runtime
+* New `pretty_urls` config option to support Nextcloud servers that require `/index.php`
+* Screenshot the active display with `nextshot --monitor`, or `-m` for short
+
+### Changed
+* Window selection on Sway is now via Slurp instead of a list
+
+### Fixed
+* Floating windows weren't able to be selected on Sway
+
+
 ## [1.3.2] - 2020-04-06
 ### Fixed
 * Printf no longer fails with non-dot decimal separators in locale
@@ -205,7 +219,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Copying link to clipboard on X11 and Wayland
 
 
-[Unreleased]: https://github.com/dshoreman/nextshot/compare/v1.3.2...develop
+[Unreleased]: https://github.com/dshoreman/nextshot/compare/v1.4.0...develop
+[1.4.0]: https://github.com/dshoreman/nextshot/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/dshoreman/nextshot/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/dshoreman/nextshot/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/dshoreman/nextshot/compare/v1.2.4...v1.3.0